// src/api-serverless/src/generated/models/ProfileMin.ts
/**
 * Seize API
 * Following is the API interface description for most common operations in Seize API. Some modifying endpoints may need authentication token.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ProfileMin {
  'id': string;
  'handle': string;
  'pfp'?: string;
  'cic': number;
  'rep': number;
  'tdh': number;
  'level': number;
  'archived': boolean;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'id',
      baseName: 'id',
      type: 'string',
      format: 'uuid'
    },
    {
      name: 'handle',
      baseName: 'handle',
      type: 'string',
      format: ''
    },
    {
      name: 'pfp',
      baseName: 'pfp',
      type: 'string',
      format: ''
    },
    {
      name: 'cic',
      baseName: 'cic',
      type: 'number',
      format: 'int64'
    },
    {
      name: 'rep',
      baseName: 'rep',
      type: 'number',
      format: 'int64'
    },
    {
      name: 'tdh',
      baseName: 'tdh',
      type: 'number',
      format: 'int64'
    },
    {
      name: 'level',
      baseName: 'level',
      type: 'number',
      format: 'int64'
    },
    {
      name: 'archived',
      baseName: 'archived',
      type: 'boolean',
      format: ''
    }
  ];

  static getAttributeTypeMap() {
    return ProfileMin.attributeTypeMap;
  }

  public constructor() {}
}